add server list component
