+++ component/server_list/app.properties
@@ -1,0 +1,7 @@
+{
+    "appId": "test",
+    "cluster": "dev",
+    "namespaceName": "application,abc1",
+    "ip": "localhost:8888",
+    "backupConfigPath":""
+}
