+++ component/server_list/app_config_server_test.go
@@ -1,4 +1,4 @@
-package component
+package server_list
 
 import (
 	"net/http"
